// src/service/request/types.ts
import { AxiosRequestConfig, AxiosResponse } from 'axios'

export interface HiRequestInterceptors<T = AxiosResponse> {
  requestInterceptor?: (config: AxiosRequestConfig) => AxiosRequestConfig
  requestInterceptorCatch?: (error: any) => any
  responseInterceptor?: (response: T) => T
  responseInterceptorCatch?: (error: any) => any
}

export interface HiRequestConfig<T = AxiosResponse> extends AxiosRequestConfig {
  interceptors?: HiRequestInterceptors<T>
}